+++ heylogs-api/src/main/java/nbbrd/heylogs/spi/ForgeRefRuleSupport.java
@@ -4,7 +4,7 @@ import com.vladsch.flexmark.ast.Link;
 import com.vladsch.flexmark.util.ast.Node;
 import lombok.NonNull;
 import nbbrd.io.text.Parser;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.util.Properties;
 import java.util.function.BiFunction;