+++ heylogs-api/src/test/java/nbbrd/heylogs/HeylogsTest.java
@@ -7,7 +7,7 @@ import internal.heylogs.base.StylishFormat;
 import lombok.NonNull;
 import nbbrd.design.MightBePromoted;
 import nbbrd.heylogs.spi.*;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;