+++ heylogs-ext-github/src/main/java/nbbrd/heylogs/ext/github/GitHubIssueRef.java
@@ -1,11 +1,11 @@
 package nbbrd.heylogs.ext.github;
 
-import nbbrd.heylogs.spi.ForgeRef;
 import lombok.AccessLevel;
 import lombok.NonNull;
 import nbbrd.design.RepresentableAsString;
 import nbbrd.design.StaticFactoryMethod;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import nbbrd.heylogs.spi.ForgeRef;
+import org.jspecify.annotations.Nullable;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;